+++ src/main/java/com/transferwise/openbanking/client/api/payment/v3/RestPaymentClient.java
@@ -1,9 +1,10 @@
 package com.transferwise.openbanking.client.api.payment.v3;
 
+import com.transferwise.openbanking.client.api.common.ApiResponse;
 import com.transferwise.openbanking.client.api.common.OpenBankingHeaders;
-import com.transferwise.openbanking.client.api.payment.common.AuthorizationContext;
-import com.transferwise.openbanking.client.api.payment.common.BasePaymentClient;
+import com.transferwise.openbanking.client.api.common.BaseClient;
 import com.transferwise.openbanking.client.api.payment.common.IdempotencyKeyGenerator;
+import com.transferwise.openbanking.client.api.payment.v3.model.OBErrorResponse1;
 import com.transferwise.openbanking.client.api.payment.v3.model.OBWriteDomestic2;
 import com.transferwise.openbanking.client.api.payment.v3.model.OBWriteDomesticConsent4;
 import com.transferwise.openbanking.client.api.payment.v3.model.OBWriteDomesticConsentResponse5;
@@ -11,10 +12,8 @@ import com.transferwise.openbanking.client.api.payment.v3.model.OBWriteDomesticR
 import com.transferwise.openbanking.client.api.payment.v3.model.OBWriteFundsConfirmationResponse1;
 import com.transferwise.openbanking.client.configuration.AspspDetails;
 import com.transferwise.openbanking.client.configuration.SoftwareStatementDetails;
-import com.transferwise.openbanking.client.error.ApiCallException;
 import com.transferwise.openbanking.client.json.JsonConverter;
 import com.transferwise.openbanking.client.jwt.JwtClaimsSigner;
-import com.transferwise.openbanking.client.oauth.OAuthClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
@@ -24,7 +23,7 @@ import org.springframework.web.client.RestClientResponseException;
 import org.springframework.web.client.RestOperations;
 
 @Slf4j
-public class RestPaymentClient extends BasePaymentClient implements PaymentClient {
+public class RestPaymentClient extends BaseClient implements PaymentClient {
 
     private static final String ENDPOINT_PATH_FORMAT = "%s/open-banking/v3.%s/pisp/%s";
 
@@ -36,21 +35,23 @@ public class RestPaymentClient extends BasePaymentClient implements PaymentClien
 
     public RestPaymentClient(RestOperations restOperations,
                              JsonConverter jsonConverter,
-                             OAuthClient oAuthClient,
                              IdempotencyKeyGenerator<OBWriteDomesticConsent4, OBWriteDomestic2> idempotencyKeyGenerator,
                              JwtClaimsSigner jwtClaimsSigner) {
-        super(restOperations, jsonConverter, oAuthClient);
+        super(restOperations, jsonConverter);
         this.idempotencyKeyGenerator = idempotencyKeyGenerator;
         this.jwtClaimsSigner = jwtClaimsSigner;
     }
 
     @Override
-    public OBWriteDomesticConsentResponse5 createDomesticPaymentConsent(OBWriteDomesticConsent4 domesticPaymentConsentRequest,
-                                                                        AspspDetails aspspDetails,
-        SoftwareStatementDetails softwareStatementDetails) {
+    public ApiResponse<OBWriteDomesticConsentResponse5, OBErrorResponse1> createDomesticPaymentConsent(
+        OBWriteDomesticConsent4 domesticPaymentConsentRequest,
+        String clientCredentialsToken,
+        AspspDetails aspspDetails,
+        SoftwareStatementDetails softwareStatementDetails
+    ) {
 
         OpenBankingHeaders headers = OpenBankingHeaders.postHeaders(aspspDetails.getOrganisationId(),
-            getClientCredentialsToken(aspspDetails),
+            clientCredentialsToken,
             idempotencyKeyGenerator.generateKeyForSetup(domesticPaymentConsentRequest),
             jwtClaimsSigner.createDetachedSignature(domesticPaymentConsentRequest,
                 aspspDetails,
@@ -62,34 +63,37 @@ public class RestPaymentClient extends BasePaymentClient implements PaymentClien
 
         log.info("Calling create payment consent API, with interaction ID {}", headers.getInteractionId());
 
-        OBWriteDomesticConsentResponse5 domesticPaymentConsentResponse;
+        ResponseEntity<String> response;
         try {
-            ResponseEntity<OBWriteDomesticConsentResponse5> response = restOperations.exchange(
-                generateApiUrl(aspspDetails, PAYMENT_CONSENT_RESOURCE),
+            response = restOperations.exchange(generateApiUrl(aspspDetails, PAYMENT_CONSENT_RESOURCE),
                 HttpMethod.POST,
                 request,
-                OBWriteDomesticConsentResponse5.class);
-            domesticPaymentConsentResponse = response.getBody();
+                String.class);
         } catch (RestClientResponseException e) {
-            throw new ApiCallException("Call to create payment consent endpoint failed, body returned '" + e.getResponseBodyAsString() + "'",
-                e);
+            return mapClientExceptionWithResponse(e, OBErrorResponse1.class);
         } catch (RestClientException e) {
-            throw new ApiCallException("Call to create payment consent endpoint failed, and no response body returned", e);
+            return mapClientException(e);
         }
 
-        validateResponse(domesticPaymentConsentResponse);
+        OBWriteDomesticConsentResponse5 domesticPaymentConsentResponse = jsonConverter.readValue(response.getBody(),
+            OBWriteDomesticConsentResponse5.class);
+        if (isResponseInvalid(domesticPaymentConsentResponse)) {
+            return mapInvalidResponse(response);
+        }
 
-        return domesticPaymentConsentResponse;
+        return ApiResponse.success(response.getStatusCodeValue(), response.getBody(), domesticPaymentConsentResponse);
     }
 
     @Override
-    public OBWriteDomesticResponse5 submitDomesticPayment(OBWriteDomestic2 domesticPaymentRequest,
-                                                          AuthorizationContext authorizationContext,
-                                                          AspspDetails aspspDetails,
-                                                          SoftwareStatementDetails softwareStatementDetails) {
+    public ApiResponse<OBWriteDomesticResponse5, OBErrorResponse1> submitDomesticPayment(
+        OBWriteDomestic2 domesticPaymentRequest,
+        String authorizationCodeToken,
+        AspspDetails aspspDetails,
+        SoftwareStatementDetails softwareStatementDetails
+    ) {
 
         OpenBankingHeaders headers = OpenBankingHeaders.postHeaders(aspspDetails.getOrganisationId(),
-            exchangeAuthorizationCode(authorizationContext, aspspDetails),
+            authorizationCodeToken,
             idempotencyKeyGenerator.generateKeyForSubmission(domesticPaymentRequest),
             jwtClaimsSigner.createDetachedSignature(domesticPaymentRequest, aspspDetails, softwareStatementDetails));
 
@@ -99,120 +103,135 @@ public class RestPaymentClient extends BasePaymentClient implements PaymentClien
 
         log.info("Calling submit payment API, with interaction ID {}", headers.getInteractionId());
 
-        OBWriteDomesticResponse5 domesticPaymentResponse;
+        ResponseEntity<String> response;
         try {
-            ResponseEntity<OBWriteDomesticResponse5> response = restOperations.exchange(
+            response = restOperations.exchange(
                 generateApiUrl(aspspDetails, PAYMENT_RESOURCE),
                 HttpMethod.POST,
                 request,
-                OBWriteDomesticResponse5.class);
-            domesticPaymentResponse = response.getBody();
+                String.class);
         } catch (RestClientResponseException e) {
-            throw new ApiCallException("Call to submit payment endpoint failed, body returned '" + e.getResponseBodyAsString() + "'",
-                e);
+            return mapClientExceptionWithResponse(e, OBErrorResponse1.class);
         } catch (RestClientException e) {
-            throw new ApiCallException("Call to submit payment endpoint failed, and no response body returned", e);
+            return mapClientException(e);
         }
 
-        validateResponse(domesticPaymentResponse);
+        OBWriteDomesticResponse5 domesticPaymentResponse = jsonConverter.readValue(response.getBody(),
+            OBWriteDomesticResponse5.class);
+        if (isResponseInvalid(domesticPaymentResponse)) {
+            return mapInvalidResponse(response);
+        }
 
-        return domesticPaymentResponse;
+        return ApiResponse.success(response.getStatusCodeValue(), response.getBody(), domesticPaymentResponse);
     }
 
     @Override
-    public OBWriteDomesticConsentResponse5 getDomesticPaymentConsent(String consentId, AspspDetails aspspDetails) {
+    public ApiResponse<OBWriteDomesticConsentResponse5, OBErrorResponse1> getDomesticPaymentConsent(
+        String consentId,
+        String clientCredentialsToken,
+        AspspDetails aspspDetails
+    ) {
 
         OpenBankingHeaders headers = OpenBankingHeaders.defaultHeaders(aspspDetails.getOrganisationId(),
-            getClientCredentialsToken(aspspDetails));
+            clientCredentialsToken);
 
         HttpEntity<?> request = new HttpEntity<>(headers);
 
         log.info("Calling get payment consent API, with interaction ID {}", headers.getInteractionId());
 
-        OBWriteDomesticConsentResponse5 domesticPaymentConsentResponse;
+        ResponseEntity<String> response;
         try {
-            ResponseEntity<OBWriteDomesticConsentResponse5> response = restOperations.exchange(
+            response = restOperations.exchange(
                 generateApiUrl(aspspDetails, PAYMENT_CONSENT_RESOURCE) + "/{consentId}",
                 HttpMethod.GET,
                 request,
-                OBWriteDomesticConsentResponse5.class,
+                String.class,
                 consentId);
-            domesticPaymentConsentResponse = response.getBody();
         } catch (RestClientResponseException e) {
-            throw new ApiCallException("Call to get payment consent endpoint failed, body returned '" + e.getResponseBodyAsString() + "'",
-                e);
+            return mapClientExceptionWithResponse(e, OBErrorResponse1.class);
         } catch (RestClientException e) {
-            throw new ApiCallException("Call to get payment consent endpoint failed, and no response body returned", e);
+            return mapClientException(e);
         }
 
-        validateResponse(domesticPaymentConsentResponse);
+        OBWriteDomesticConsentResponse5 domesticPaymentConsentResponse = jsonConverter.readValue(response.getBody(),
+            OBWriteDomesticConsentResponse5.class);
+        if (isResponseInvalid(domesticPaymentConsentResponse)) {
+            return mapInvalidResponse(response);
+        }
 
-        return domesticPaymentConsentResponse;
+        return ApiResponse.success(response.getStatusCodeValue(), response.getBody(), domesticPaymentConsentResponse);
     }
 
     @Override
-    public OBWriteDomesticResponse5 getDomesticPayment(String domesticPaymentId, AspspDetails aspspDetails) {
+    public ApiResponse<OBWriteDomesticResponse5, OBErrorResponse1> getDomesticPayment(String domesticPaymentId,
+                                                                                      String clientCredentialsToken,
+                                                                                      AspspDetails aspspDetails) {
 
         OpenBankingHeaders headers = OpenBankingHeaders.defaultHeaders(aspspDetails.getOrganisationId(),
-            getClientCredentialsToken(aspspDetails));
+            clientCredentialsToken);
 
         HttpEntity<?> request = new HttpEntity<>(headers);
 
         log.info("Calling get payment API, with interaction ID {}", headers.getInteractionId());
 
-        OBWriteDomesticResponse5 domesticPaymentResponse;
+        ResponseEntity<String> response;
         try {
-            ResponseEntity<OBWriteDomesticResponse5> response = restOperations.exchange(
+            response = restOperations.exchange(
                 generateApiUrl(aspspDetails, PAYMENT_RESOURCE) + "/{domesticPaymentId}",
                 HttpMethod.GET,
                 request,
-                OBWriteDomesticResponse5.class,
+                String.class,
                 domesticPaymentId);
-            domesticPaymentResponse = response.getBody();
         } catch (RestClientResponseException e) {
-            throw new ApiCallException("Call to get payment endpoint failed, body returned '" + e.getResponseBodyAsString() + "'",
-                e);
+            return mapClientExceptionWithResponse(e, OBErrorResponse1.class);
         } catch (RestClientException e) {
-            throw new ApiCallException("Call to get payment endpoint failed, and no response body returned", e);
+            return mapClientException(e);
         }
 
-        validateResponse(domesticPaymentResponse);
+        OBWriteDomesticResponse5 domesticPaymentResponse = jsonConverter.readValue(response.getBody(),
+            OBWriteDomesticResponse5.class);
+        if (isResponseInvalid(domesticPaymentResponse)) {
+            return mapInvalidResponse(response);
+        }
 
-        return domesticPaymentResponse;
+        return ApiResponse.success(response.getStatusCodeValue(), response.getBody(), domesticPaymentResponse);
     }
 
     @Override
-    public OBWriteFundsConfirmationResponse1 getFundsConfirmation(String consentId,
-                                                                  AuthorizationContext authorizationContext,
-                                                                  AspspDetails aspspDetails) {
+    public ApiResponse<OBWriteFundsConfirmationResponse1, OBErrorResponse1> getFundsConfirmation(
+        String consentId,
+        String authorizationCodeToken,
+        AspspDetails aspspDetails
+    ) {
 
         OpenBankingHeaders headers = OpenBankingHeaders.defaultHeaders(aspspDetails.getOrganisationId(),
-            exchangeAuthorizationCode(authorizationContext, aspspDetails));
+            authorizationCodeToken);
 
         HttpEntity<?> request = new HttpEntity<>(headers);
 
         log.info("Calling get confirmation of funds API, with interaction ID {}", headers.getInteractionId());
 
-        OBWriteFundsConfirmationResponse1 fundsConfirmationResponse;
+        ResponseEntity<String> response;
         try {
-            ResponseEntity<OBWriteFundsConfirmationResponse1> response = restOperations.exchange(
+            response = restOperations.exchange(
                 generateApiUrl(aspspDetails, PAYMENT_CONSENT_RESOURCE) + "/{consentId}/funds-confirmation",
                 HttpMethod.GET,
                 request,
-                OBWriteFundsConfirmationResponse1.class,
+                String.class,
                 consentId);
-            fundsConfirmationResponse = response.getBody();
         } catch (RestClientResponseException e) {
-            throw new ApiCallException("Call to get confirmation of funds endpoint failed, body returned '" + e.getResponseBodyAsString() + "'",
-                e);
+            return mapClientExceptionWithResponse(e, OBErrorResponse1.class);
         } catch (RestClientException e) {
-            throw new ApiCallException("Call to get confirmation of funds endpoint failed, and no response body returned",
-                e);
+            return mapClientException(e);
         }
 
-        validateResponse(fundsConfirmationResponse);
+        OBWriteFundsConfirmationResponse1 fundsConfirmationResponse = jsonConverter.readValue(response.getBody(),
+            OBWriteFundsConfirmationResponse1.class);
+        if (isResponseInvalid(fundsConfirmationResponse)) {
+            return mapInvalidResponse(response);
+        }
 
-        return fundsConfirmationResponse;
+        return ApiResponse.success(response.getStatusCodeValue(), response.getBody(), fundsConfirmationResponse);
     }
 
     private String generateApiUrl(AspspDetails aspspDetails, String resource) {
@@ -222,29 +241,23 @@ public class RestPaymentClient extends BasePaymentClient implements PaymentClien
             resource);
     }
 
-    private void validateResponse(OBWriteDomesticConsentResponse5 response) {
-        if (response == null ||
+    private boolean isResponseInvalid(OBWriteDomesticConsentResponse5 response) {
+        return response == null ||
             response.getData() == null ||
             response.getData().getStatus() == null ||
             response.getData().getConsentId() == null ||
-            response.getData().getConsentId().isBlank()) {
-            throw new ApiCallException("Empty or partial domestic payment consent response returned " + response);
-        }
+            response.getData().getConsentId().isBlank();
     }
 
-    private void validateResponse(OBWriteDomesticResponse5 response) {
-        if (response == null ||
+    private boolean isResponseInvalid(OBWriteDomesticResponse5 response) {
+        return response == null ||
             response.getData() == null ||
             response.getData().getStatus() == null ||
             response.getData().getDomesticPaymentId() == null ||
-            response.getData().getDomesticPaymentId().isBlank()) {
-            throw new ApiCallException("Empty or partial domestic payment response returned " + response);
-        }
+            response.getData().getDomesticPaymentId().isBlank();
     }
 
-    private void validateResponse(OBWriteFundsConfirmationResponse1 response) {
-        if (response == null || response.getData() == null) {
-            throw new ApiCallException("Empty or partial funds confirmation response returned " + response);
-        }
+    private boolean isResponseInvalid(OBWriteFundsConfirmationResponse1 response) {
+        return response == null || response.getData() == null;
     }
 }