+++ src/test/java/com/transferwise/openbanking/client/test/TestAspspDetails.java
@@ -1,5 +1,6 @@
 package com.transferwise.openbanking.client.test;
 
+import com.transferwise.openbanking.client.oauth.domain.Scope;
 import com.transferwise.openbanking.client.configuration.AspspDetails;
 import com.transferwise.openbanking.client.configuration.SoftwareStatementDetails;
 import com.transferwise.openbanking.client.oauth.ClientAuthenticationMethod;
@@ -9,6 +10,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.util.List;
+import java.util.Set;
 
 @Data
 @Builder
@@ -20,6 +22,7 @@ public class TestAspspDetails implements AspspDetails {
     private String registrationUrl;
     private String registrationAudience;
     private String registrationIssuer;
+    private Set<Scope> registrationAuthenticationScopes;
     private ClientAuthenticationMethod clientAuthenticationMethod;
     private String clientId;
     private String clientSecret;
@@ -32,7 +35,6 @@ public class TestAspspDetails implements AspspDetails {
     private boolean registrationUsesJoseContentType;
     private boolean detachedSignatureUsesDirectoryIssFormat;
     private boolean registrationRequiresLowerCaseJtiClaim;
-    private boolean registrationAuthenticationRequiresOpenIdScope;
 
     @Override
     public String getApiBaseUrl(String majorVersion, String resource) {
@@ -42,6 +44,11 @@ public class TestAspspDetails implements AspspDetails {
     @Override
     public String getRegistrationIssuer(SoftwareStatementDetails softwareStatementDetails) {
         return registrationIssuer;
+    }
+
+    @Override
+    public Set<Scope> getRegistrationAuthenticationScopes(SoftwareStatementDetails softwareStatementDetails) {
+        return registrationAuthenticationScopes;
     }
 
     @Override
@@ -57,10 +64,5 @@ public class TestAspspDetails implements AspspDetails {
     @Override
     public boolean registrationRequiresLowerCaseJtiClaim() {
         return registrationRequiresLowerCaseJtiClaim;
-    }
-
-    @Override
-    public boolean registrationAuthenticationRequiresOpenIdScope() {
-        return registrationAuthenticationRequiresOpenIdScope;
     }
 }